// packages/wrangler/src/commands/pipelines/types.ts
export type JsonFormat = {
	type: "json";
};

export type ParquetFormat = {
	type: "parquet";
	compression?: "uncompressed" | "snappy" | "gzip" | "zstd" | "lz4";
	row_group_bytes?: number;
};

export type SinkFormat = JsonFormat | ParquetFormat;

// Schema types
export type SchemaField = {
	name: string;
	type:
		| "bool"
		| "int32"
		| "int64"
		| "float32"
		| "float64"
		| "string"
		| "timestamp"
		| "json"
		| "bytes"
		| "list"
		| "struct";
	required: boolean;
	fields?: SchemaField[];
	items?: SchemaField;
	unit?: "second" | "millisecond" | "microsecond" | "nanosecond"; // For timestamp type
};

export type Sink = {
	id: string;
	name: string;
	created_at?: string;
	modified_at?: string;
	version?: number;
	type: "r2" | "r2_data_catalog";
	format: SinkFormat;
	schema: {
		fields: SchemaField[];
	} | null;
	config: {
		bucket: string;
		path?: string;
		partitioning?: {
			time_pattern: string;
		};
		rolling_policy?: {
			file_size_bytes?: number;
			interval_seconds: number;
		};
		// r2_data_catalog specific fields
		account_id?: string;
		token?: string;
		namespace?: string;
		table_name?: string;
	};
	used_by?: Array<{
		href: string;
	}>;
};